core/api/requirements.txt
ffmpeg-python>=0.2.0
requests>=2.31
Flask>=3.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
Flask-SocketIO>=5.4
debugpy>=1.8
eventlet>=0.36
Pillow>=10.0
